fix: move _apply_session_type_visibility out of __init__.py + fix docstring count
- Extract function to server/_session_type.py to satisfy pure-facade constraint
  (test_init_files_are_pure_facades: no FunctionDef at module scope in __init__.py)
- Re-export via server/__init__.py import so test imports remain unchanged
- Update docstring to include '41 kitchen-tagged' substring for docstring count test
- Use deferred import of mcp inside function to avoid circular import at module level
